Update spring codegen composer tests to use golden files
