+++ BUILD.bazel
@@ -208,6 +208,8 @@ GOLDEN_UPDATING_UNIT_TESTS = [
     "com.google.api.generator.gapic.composer.rest.ServiceClientTestClassComposerTest",
     "com.google.api.generator.gapic.composer.rest.ServiceSettingsClassComposerTest",
     "com.google.api.generator.gapic.composer.rest.ServiceStubSettingsClassComposerTest",
+    "com.google.api.generator.spring.SpringAutoConfigClassComposerTest",
+    "com.google.api.generator.spring.SpringPropertiesClassComposerTest",
 ]
 
 # Manually runs a single JUnit test for the tests that generate and save .golden
