+++ src/test/java/com/google/api/generator/spring/SpringAutoConfigClassComposerTest.java
@@ -29,7 +29,8 @@ import com.google.api.generator.gapic.model.Transport;
 import com.google.api.generator.gapic.protoparser.Parser;
 import com.google.api.generator.gapic.protoparser.ServiceConfigParser;
 import com.google.api.generator.spring.composer.SpringAutoConfigClassComposer;
-import com.google.api.generator.spring.composer.SpringPropertiesClassComposer;
+import com.google.api.generator.test.framework.Assert;
+import com.google.api.generator.test.framework.Utils;
 import com.google.protobuf.Descriptors.FileDescriptor;
 import com.google.protobuf.Descriptors.ServiceDescriptor;
 import com.google.showcase.v1beta1.EchoOuterClass;
@@ -89,47 +90,13 @@ public class SpringAutoConfigClassComposerTest {
 
     JavaWriterVisitor visitor = new JavaWriterVisitor();
     clazz.classDefinition().accept(visitor);
-    assertEquals(EXPECTED_AUTOCONFIG_CLASS_STRING, visitor.write());
+    Utils.saveCodegenToFile(this.getClass(), "SpringAutoConfigClass.golden", visitor.write());
+    Path goldenFilePath =
+        Paths.get(Utils.getGoldenDir(this.getClass()), "SpringAutoConfigClass.golden");
+    Assert.assertCodeEquals(goldenFilePath, visitor.write());
   }
 
-  @Test
-  public void generatePropertiesTest() {
-
-    Map<String, Message> messageTypes = Parser.parseMessages(echoFileDescriptor);
-
-    Map<String, ResourceName> resourceNames = Parser.parseResourceNames(echoFileDescriptor);
-    Set<ResourceName> outputResourceNames = new HashSet<>();
-
-    List<Service> services =
-        Parser.parseService(
-            echoFileDescriptor, messageTypes, resourceNames, Optional.empty(), outputResourceNames);
-
-    String jsonFilename = "showcase_grpc_service_config.json";
-    Path jsonPath = Paths.get(TestProtoLoader.instance().getTestFilesDirectory(), jsonFilename);
-    Optional<GapicServiceConfig> serviceConfigOpt = ServiceConfigParser.parse(jsonPath.toString());
-    assertTrue(serviceConfigOpt.isPresent());
-    GapicServiceConfig serviceConfig = serviceConfigOpt.get();
-
-    GapicContext context =
-        GapicContext.builder()
-            .setMessages(messageTypes)
-            .setResourceNames(resourceNames)
-            .setServices(services)
-            .setHelperResourceNames(outputResourceNames)
-            .setTransport(Transport.GRPC)
-            .setServiceConfig(serviceConfig)
-            .build();
-
-    Service echoProtoService = services.get(0);
-
-    GapicClass clazz = SpringPropertiesClassComposer.instance().generate(context, echoProtoService);
-
-    JavaWriterVisitor visitor = new JavaWriterVisitor();
-    clazz.classDefinition().accept(visitor);
-    System.out.println(visitor.write());
-    assertEquals(EXPECTED_PROPERTY_CLASS_STRING, visitor.write());
-  }
-
+  // TODO: Remove this string once SpringComposer is refactored
   protected static final String EXPECTED_AUTOCONFIG_CLASS_STRING =
       "package com.google.showcase.v1beta1.spring;\n"
           + "\n"
@@ -290,493 +257,6 @@ public class SpringAutoConfigClassComposerTest {
           + "        .collideNameSettings()\n"
           + "        .setRetrySettings(collideNameRetrySettingBuilder.build());\n"
           + "    return EchoClient.create(clientSettingsBuilder.build());\n"
-          + "  }\n"
-          + "}\n";
-
-  protected static final String EXPECTED_PROPERTY_CLASS_STRING =
-      "package com.google.showcase.v1beta1.spring;\n"
-          + "\n"
-          + "import com.google.cloud.spring.core.Credentials;\n"
-          + "import com.google.cloud.spring.core.CredentialsSupplier;\n"
-          + "import org.springframework.boot.context.properties.ConfigurationProperties;\n"
-          + "import org.threeten.bp.Duration;\n"
-          + "\n"
-          + "@ConfigurationProperties(\"spring.cloud.gcp.autoconfig.showcase.echo\")\n"
-          + "public class EchoSpringProperties implements CredentialsSupplier {\n"
-          + "  private final Credentials credentials =\n"
-          + "      new Credentials(\"https://www.googleapis.com/auth/cloud-platform\");\n"
-          + "  private String quotaProjectId;\n"
-          + "  private Integer executorThreadCount;\n"
-          + "  private boolean useRest;\n"
-          // retry settings per method, no default set
-          // checks for null in autoconfig class when using
-          + "  private Duration echoInitialRetryDelay;\n"
-          + "  private Double echoRetryDelayMultiplier;\n"
-          + "  private Duration echoMaxRetryDelay;\n"
-          + "  private Duration echoInitialRpcTimeout;\n"
-          + "  private Double echoRpcTimeoutMultiplier;\n"
-          + "  private Duration echoMaxRpcTimeout;\n"
-          + "  private Duration echoTotalTimeout;\n"
-          + "  private Duration expandInitialRetryDelay;\n"
-          + "  private Double expandRetryDelayMultiplier;\n"
-          + "  private Duration expandMaxRetryDelay;\n"
-          + "  private Duration expandInitialRpcTimeout;\n"
-          + "  private Double expandRpcTimeoutMultiplier;\n"
-          + "  private Duration expandMaxRpcTimeout;\n"
-          + "  private Duration expandTotalTimeout;\n"
-          + "  private Duration collectInitialRpcTimeout;\n"
-          + "  private Double collectRpcTimeoutMultiplier;\n"
-          + "  private Duration collectMaxRpcTimeout;\n"
-          + "  private Duration collectTotalTimeout;\n"
-          + "  private Duration chatInitialRpcTimeout;\n"
-          + "  private Double chatRpcTimeoutMultiplier;\n"
-          + "  private Duration chatMaxRpcTimeout;\n"
-          + "  private Duration chatTotalTimeout;\n"
-          + "  private Duration chatAgainInitialRpcTimeout;\n"
-          + "  private Double chatAgainRpcTimeoutMultiplier;\n"
-          + "  private Duration chatAgainMaxRpcTimeout;\n"
-          + "  private Duration chatAgainTotalTimeout;\n"
-          + "  private Duration pagedExpandInitialRetryDelay;\n"
-          + "  private Double pagedExpandRetryDelayMultiplier;\n"
-          + "  private Duration pagedExpandMaxRetryDelay;\n"
-          + "  private Duration pagedExpandInitialRpcTimeout;\n"
-          + "  private Double pagedExpandRpcTimeoutMultiplier;\n"
-          + "  private Duration pagedExpandMaxRpcTimeout;\n"
-          + "  private Duration pagedExpandTotalTimeout;\n"
-          + "  private Duration simplePagedExpandInitialRpcTimeout;\n"
-          + "  private Double simplePagedExpandRpcTimeoutMultiplier;\n"
-          + "  private Duration simplePagedExpandMaxRpcTimeout;\n"
-          + "  private Duration simplePagedExpandTotalTimeout;\n"
-          + "  private Duration waitInitialRpcTimeout;\n"
-          + "  private Double waitRpcTimeoutMultiplier;\n"
-          + "  private Duration waitMaxRpcTimeout;\n"
-          + "  private Duration waitTotalTimeout;\n"
-          + "  private Duration blockInitialRpcTimeout;\n"
-          + "  private Double blockRpcTimeoutMultiplier;\n"
-          + "  private Duration blockMaxRpcTimeout;\n"
-          + "  private Duration blockTotalTimeout;\n"
-          + "  private Duration collideNameInitialRpcTimeout;\n"
-          + "  private Double collideNameRpcTimeoutMultiplier;\n"
-          + "  private Duration collideNameMaxRpcTimeout;\n"
-          + "  private Duration collideNameTotalTimeout;\n"
-          + "\n"
-          + "  @Override\n"
-          + "  public Credentials getCredentials() {\n"
-          + "    return this.credentials;\n"
-          + "  }\n"
-          + "\n"
-          + "  public String getQuotaProjectId() {\n"
-          + "    return this.quotaProjectId;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setQuotaProjectId(String quotaProjectId) {\n"
-          + "    this.quotaProjectId = quotaProjectId;\n"
-          + "  }\n"
-          + "\n"
-          + "  public boolean getUseRest() {\n"
-          + "    return this.useRest;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Integer getExecutorThreadCount() {\n"
-          + "    return this.executorThreadCount;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setExecutorThreadCount(Integer executorThreadCount) {\n"
-          + "    this.executorThreadCount = executorThreadCount;\n"
-          + "  }\n"
-          + "\n"
-          // getters and setters for retry settings.
-          + "  public Duration getEchoInitialRetryDelay() {\n"
-          + "    return this.echoInitialRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setEchoInitialRetryDelay(Duration echoInitialRetryDelay) {\n"
-          + "    this.echoInitialRetryDelay = echoInitialRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getEchoRetryDelayMultiplier() {\n"
-          + "    return this.echoRetryDelayMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setEchoRetryDelayMultiplier(Double echoRetryDelayMultiplier) {\n"
-          + "    this.echoRetryDelayMultiplier = echoRetryDelayMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getEchoMaxRetryDelay() {\n"
-          + "    return this.echoMaxRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setEchoMaxRetryDelay(Duration echoMaxRetryDelay) {\n"
-          + "    this.echoMaxRetryDelay = echoMaxRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getEchoInitialRpcTimeout() {\n"
-          + "    return this.echoInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setEchoInitialRpcTimeout(Duration echoInitialRpcTimeout) {\n"
-          + "    this.echoInitialRpcTimeout = echoInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getEchoRpcTimeoutMultiplier() {\n"
-          + "    return this.echoRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setEchoRpcTimeoutMultiplier(Double echoRpcTimeoutMultiplier) {\n"
-          + "    this.echoRpcTimeoutMultiplier = echoRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getEchoMaxRpcTimeout() {\n"
-          + "    return this.echoMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setEchoMaxRpcTimeout(Duration echoMaxRpcTimeout) {\n"
-          + "    this.echoMaxRpcTimeout = echoMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getEchoTotalTimeout() {\n"
-          + "    return this.echoTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setEchoTotalTimeout(Duration echoTotalTimeout) {\n"
-          + "    this.echoTotalTimeout = echoTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getExpandInitialRetryDelay() {\n"
-          + "    return this.expandInitialRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setExpandInitialRetryDelay(Duration expandInitialRetryDelay) {\n"
-          + "    this.expandInitialRetryDelay = expandInitialRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getExpandRetryDelayMultiplier() {\n"
-          + "    return this.expandRetryDelayMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setExpandRetryDelayMultiplier(Double expandRetryDelayMultiplier) {\n"
-          + "    this.expandRetryDelayMultiplier = expandRetryDelayMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getExpandMaxRetryDelay() {\n"
-          + "    return this.expandMaxRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setExpandMaxRetryDelay(Duration expandMaxRetryDelay) {\n"
-          + "    this.expandMaxRetryDelay = expandMaxRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getExpandInitialRpcTimeout() {\n"
-          + "    return this.expandInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setExpandInitialRpcTimeout(Duration expandInitialRpcTimeout) {\n"
-          + "    this.expandInitialRpcTimeout = expandInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getExpandRpcTimeoutMultiplier() {\n"
-          + "    return this.expandRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setExpandRpcTimeoutMultiplier(Double expandRpcTimeoutMultiplier) {\n"
-          + "    this.expandRpcTimeoutMultiplier = expandRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getExpandMaxRpcTimeout() {\n"
-          + "    return this.expandMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setExpandMaxRpcTimeout(Duration expandMaxRpcTimeout) {\n"
-          + "    this.expandMaxRpcTimeout = expandMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getExpandTotalTimeout() {\n"
-          + "    return this.expandTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setExpandTotalTimeout(Duration expandTotalTimeout) {\n"
-          + "    this.expandTotalTimeout = expandTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getCollectInitialRpcTimeout() {\n"
-          + "    return this.collectInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setCollectInitialRpcTimeout(Duration collectInitialRpcTimeout) {\n"
-          + "    this.collectInitialRpcTimeout = collectInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getCollectRpcTimeoutMultiplier() {\n"
-          + "    return this.collectRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setCollectRpcTimeoutMultiplier(Double collectRpcTimeoutMultiplier) {\n"
-          + "    this.collectRpcTimeoutMultiplier = collectRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getCollectMaxRpcTimeout() {\n"
-          + "    return this.collectMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setCollectMaxRpcTimeout(Duration collectMaxRpcTimeout) {\n"
-          + "    this.collectMaxRpcTimeout = collectMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getCollectTotalTimeout() {\n"
-          + "    return this.collectTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setCollectTotalTimeout(Duration collectTotalTimeout) {\n"
-          + "    this.collectTotalTimeout = collectTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getChatInitialRpcTimeout() {\n"
-          + "    return this.chatInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setChatInitialRpcTimeout(Duration chatInitialRpcTimeout) {\n"
-          + "    this.chatInitialRpcTimeout = chatInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getChatRpcTimeoutMultiplier() {\n"
-          + "    return this.chatRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setChatRpcTimeoutMultiplier(Double chatRpcTimeoutMultiplier) {\n"
-          + "    this.chatRpcTimeoutMultiplier = chatRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getChatMaxRpcTimeout() {\n"
-          + "    return this.chatMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setChatMaxRpcTimeout(Duration chatMaxRpcTimeout) {\n"
-          + "    this.chatMaxRpcTimeout = chatMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getChatTotalTimeout() {\n"
-          + "    return this.chatTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setChatTotalTimeout(Duration chatTotalTimeout) {\n"
-          + "    this.chatTotalTimeout = chatTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getChatAgainInitialRpcTimeout() {\n"
-          + "    return this.chatAgainInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setChatAgainInitialRpcTimeout(Duration chatAgainInitialRpcTimeout) {\n"
-          + "    this.chatAgainInitialRpcTimeout = chatAgainInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getChatAgainRpcTimeoutMultiplier() {\n"
-          + "    return this.chatAgainRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setChatAgainRpcTimeoutMultiplier(Double chatAgainRpcTimeoutMultiplier) {\n"
-          + "    this.chatAgainRpcTimeoutMultiplier = chatAgainRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getChatAgainMaxRpcTimeout() {\n"
-          + "    return this.chatAgainMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setChatAgainMaxRpcTimeout(Duration chatAgainMaxRpcTimeout) {\n"
-          + "    this.chatAgainMaxRpcTimeout = chatAgainMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getChatAgainTotalTimeout() {\n"
-          + "    return this.chatAgainTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setChatAgainTotalTimeout(Duration chatAgainTotalTimeout) {\n"
-          + "    this.chatAgainTotalTimeout = chatAgainTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getPagedExpandInitialRetryDelay() {\n"
-          + "    return this.pagedExpandInitialRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setPagedExpandInitialRetryDelay(Duration pagedExpandInitialRetryDelay) {\n"
-          + "    this.pagedExpandInitialRetryDelay = pagedExpandInitialRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getPagedExpandRetryDelayMultiplier() {\n"
-          + "    return this.pagedExpandRetryDelayMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setPagedExpandRetryDelayMultiplier(Double pagedExpandRetryDelayMultiplier) {\n"
-          + "    this.pagedExpandRetryDelayMultiplier = pagedExpandRetryDelayMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getPagedExpandMaxRetryDelay() {\n"
-          + "    return this.pagedExpandMaxRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setPagedExpandMaxRetryDelay(Duration pagedExpandMaxRetryDelay) {\n"
-          + "    this.pagedExpandMaxRetryDelay = pagedExpandMaxRetryDelay;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getPagedExpandInitialRpcTimeout() {\n"
-          + "    return this.pagedExpandInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setPagedExpandInitialRpcTimeout(Duration pagedExpandInitialRpcTimeout) {\n"
-          + "    this.pagedExpandInitialRpcTimeout = pagedExpandInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getPagedExpandRpcTimeoutMultiplier() {\n"
-          + "    return this.pagedExpandRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setPagedExpandRpcTimeoutMultiplier(Double pagedExpandRpcTimeoutMultiplier) {\n"
-          + "    this.pagedExpandRpcTimeoutMultiplier = pagedExpandRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getPagedExpandMaxRpcTimeout() {\n"
-          + "    return this.pagedExpandMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setPagedExpandMaxRpcTimeout(Duration pagedExpandMaxRpcTimeout) {\n"
-          + "    this.pagedExpandMaxRpcTimeout = pagedExpandMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getPagedExpandTotalTimeout() {\n"
-          + "    return this.pagedExpandTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setPagedExpandTotalTimeout(Duration pagedExpandTotalTimeout) {\n"
-          + "    this.pagedExpandTotalTimeout = pagedExpandTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getSimplePagedExpandInitialRpcTimeout() {\n"
-          + "    return this.simplePagedExpandInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setSimplePagedExpandInitialRpcTimeout(Duration simplePagedExpandInitialRpcTimeout) {\n"
-          + "    this.simplePagedExpandInitialRpcTimeout = simplePagedExpandInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getSimplePagedExpandRpcTimeoutMultiplier() {\n"
-          + "    return this.simplePagedExpandRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setSimplePagedExpandRpcTimeoutMultiplier(\n"
-          + "      Double simplePagedExpandRpcTimeoutMultiplier) {\n"
-          + "    this.simplePagedExpandRpcTimeoutMultiplier = simplePagedExpandRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getSimplePagedExpandMaxRpcTimeout() {\n"
-          + "    return this.simplePagedExpandMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setSimplePagedExpandMaxRpcTimeout(Duration simplePagedExpandMaxRpcTimeout) {\n"
-          + "    this.simplePagedExpandMaxRpcTimeout = simplePagedExpandMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getSimplePagedExpandTotalTimeout() {\n"
-          + "    return this.simplePagedExpandTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setSimplePagedExpandTotalTimeout(Duration simplePagedExpandTotalTimeout) {\n"
-          + "    this.simplePagedExpandTotalTimeout = simplePagedExpandTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getWaitInitialRpcTimeout() {\n"
-          + "    return this.waitInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setWaitInitialRpcTimeout(Duration waitInitialRpcTimeout) {\n"
-          + "    this.waitInitialRpcTimeout = waitInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getWaitRpcTimeoutMultiplier() {\n"
-          + "    return this.waitRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setWaitRpcTimeoutMultiplier(Double waitRpcTimeoutMultiplier) {\n"
-          + "    this.waitRpcTimeoutMultiplier = waitRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getWaitMaxRpcTimeout() {\n"
-          + "    return this.waitMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setWaitMaxRpcTimeout(Duration waitMaxRpcTimeout) {\n"
-          + "    this.waitMaxRpcTimeout = waitMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getWaitTotalTimeout() {\n"
-          + "    return this.waitTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setWaitTotalTimeout(Duration waitTotalTimeout) {\n"
-          + "    this.waitTotalTimeout = waitTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getBlockInitialRpcTimeout() {\n"
-          + "    return this.blockInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setBlockInitialRpcTimeout(Duration blockInitialRpcTimeout) {\n"
-          + "    this.blockInitialRpcTimeout = blockInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getBlockRpcTimeoutMultiplier() {\n"
-          + "    return this.blockRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setBlockRpcTimeoutMultiplier(Double blockRpcTimeoutMultiplier) {\n"
-          + "    this.blockRpcTimeoutMultiplier = blockRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getBlockMaxRpcTimeout() {\n"
-          + "    return this.blockMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setBlockMaxRpcTimeout(Duration blockMaxRpcTimeout) {\n"
-          + "    this.blockMaxRpcTimeout = blockMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getBlockTotalTimeout() {\n"
-          + "    return this.blockTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setBlockTotalTimeout(Duration blockTotalTimeout) {\n"
-          + "    this.blockTotalTimeout = blockTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getCollideNameInitialRpcTimeout() {\n"
-          + "    return this.collideNameInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setCollideNameInitialRpcTimeout(Duration collideNameInitialRpcTimeout) {\n"
-          + "    this.collideNameInitialRpcTimeout = collideNameInitialRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Double getCollideNameRpcTimeoutMultiplier() {\n"
-          + "    return this.collideNameRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setCollideNameRpcTimeoutMultiplier(Double collideNameRpcTimeoutMultiplier) {\n"
-          + "    this.collideNameRpcTimeoutMultiplier = collideNameRpcTimeoutMultiplier;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getCollideNameMaxRpcTimeout() {\n"
-          + "    return this.collideNameMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setCollideNameMaxRpcTimeout(Duration collideNameMaxRpcTimeout) {\n"
-          + "    this.collideNameMaxRpcTimeout = collideNameMaxRpcTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public Duration getCollideNameTotalTimeout() {\n"
-          + "    return this.collideNameTotalTimeout;\n"
-          + "  }\n"
-          + "\n"
-          + "  public void setCollideNameTotalTimeout(Duration collideNameTotalTimeout) {\n"
-          + "    this.collideNameTotalTimeout = collideNameTotalTimeout;\n"
           + "  }\n"
           + "}\n";
 }
