+++ src/test/java/com/google/api/generator/spring/SpringPropertiesClassComposerTest.java
@@ -1,0 +1,101 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.google.api.generator.spring;
+
+import static junit.framework.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
+import com.google.api.generator.engine.writer.JavaWriterVisitor;
+import com.google.api.generator.gapic.composer.common.TestProtoLoader;
+import com.google.api.generator.gapic.model.GapicClass;
+import com.google.api.generator.gapic.model.GapicContext;
+import com.google.api.generator.gapic.model.GapicServiceConfig;
+import com.google.api.generator.gapic.model.Message;
+import com.google.api.generator.gapic.model.ResourceName;
+import com.google.api.generator.gapic.model.Service;
+import com.google.api.generator.gapic.model.Transport;
+import com.google.api.generator.gapic.protoparser.Parser;
+import com.google.api.generator.gapic.protoparser.ServiceConfigParser;
+import com.google.api.generator.spring.composer.SpringPropertiesClassComposer;
+import com.google.api.generator.test.framework.Assert;
+import com.google.api.generator.test.framework.Utils;
+import com.google.protobuf.Descriptors.FileDescriptor;
+import com.google.protobuf.Descriptors.ServiceDescriptor;
+import com.google.showcase.v1beta1.EchoOuterClass;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import org.junit.Before;
+import org.junit.Test;
+
+public class SpringPropertiesClassComposerTest {
+  private ServiceDescriptor echoService;
+  private FileDescriptor echoFileDescriptor;
+
+  @Before
+  public void setUp() {
+    echoFileDescriptor = EchoOuterClass.getDescriptor();
+
+    ServiceDescriptor serviceDescriptor = echoFileDescriptor.getServices().get(0);
+    // Assert.assertEquals(serviceDescriptor.getName(), "Bookshop");
+    echoService = echoFileDescriptor.getServices().get(0);
+    assertEquals(echoService.getName(), "Echo");
+  }
+
+  @Test
+  public void generatePropertiesTest() {
+
+    Map<String, Message> messageTypes = Parser.parseMessages(echoFileDescriptor);
+
+    Map<String, ResourceName> resourceNames = Parser.parseResourceNames(echoFileDescriptor);
+    Set<ResourceName> outputResourceNames = new HashSet<>();
+
+    List<Service> services =
+        Parser.parseService(
+            echoFileDescriptor, messageTypes, resourceNames, Optional.empty(), outputResourceNames);
+
+    String jsonFilename = "showcase_grpc_service_config.json";
+    Path jsonPath = Paths.get(TestProtoLoader.instance().getTestFilesDirectory(), jsonFilename);
+    Optional<GapicServiceConfig> serviceConfigOpt = ServiceConfigParser.parse(jsonPath.toString());
+    assertTrue(serviceConfigOpt.isPresent());
+    GapicServiceConfig serviceConfig = serviceConfigOpt.get();
+
+    GapicContext context =
+        GapicContext.builder()
+            .setMessages(messageTypes)
+            .setResourceNames(resourceNames)
+            .setServices(services)
+            .setHelperResourceNames(outputResourceNames)
+            .setTransport(Transport.GRPC)
+            .setServiceConfig(serviceConfig)
+            .build();
+
+    Service echoProtoService = services.get(0);
+
+    GapicClass clazz = SpringPropertiesClassComposer.instance().generate(context, echoProtoService);
+
+    JavaWriterVisitor visitor = new JavaWriterVisitor();
+    clazz.classDefinition().accept(visitor);
+    // System.out.println(visitor.write());
+    Utils.saveCodegenToFile(this.getClass(), "SpringPropertiesClass.golden", visitor.write());
+    Path goldenFilePath =
+        Paths.get(Utils.getGoldenDir(this.getClass()), "SpringPropertiesClass.golden");
+    Assert.assertCodeEquals(goldenFilePath, visitor.write());
+  }
+}
